==> Raiyan_code/1c_compile_preprocessed_corpus.ipynb <==
import numpy as np
import pandas as pd 
import json
from collections import defaultdict
import copy
import pickle 

# Load data

f = open("data/obj_all_abstracts_year.json")
data = json.load(f)

f2 = open("data/processed_abstracts.json")
abstracts_list = json.load(f2)

# Streamline data 

# these are the ids in processed$docs.removed and out$docs.removed during preproecessing in R
docs_removed = [4920,10229,16848,17604,24092,68898,104301,108465,109275,118546,118907,128414,132047,135300,
                157901,159166,159813,170490,184016,191411,195803,203451,204141,214446,221736,223792,241881,
                246810,88613,126377] 

updated_abstracts_list = []
for idx,abstract in enumerate(abstracts_list):
    if idx+1 not in docs_removed:
        updated_abstracts_list.append(abstract)

# vocab_id to actual word mapping
id_to_word = defaultdict()
word_to_id = defaultdict()
for idx,word in enumerate(data['vocab']):
    id_to_word[idx] = word
    word_to_id[word] = idx

doc_info_dict_list = [] 

doc_id = 0
for doc_id_str, doc_list in data['documents'].items():
    # get word count distribution
    temp_word_count_dict = defaultdict()
    for id_,count_ in zip(doc_list[0],doc_list[1]):
        temp_word_count_dict[id_to_word[id_-1]]=count_ # -1 to compensate for R being 1-indexed

    temp_info_dict = defaultdict()
    temp_info_dict['abstract_word_distr'] = copy.deepcopy(temp_word_count_dict)
    
    # get updated abstract
    temp_info_dict['updated_abstract'] = updated_abstracts_list[doc_id]
    
    # get other doc info
    temp_info_dict['full_abstract'] = data['meta'][doc_id]['abstract']
    temp_info_dict['doi'] = data['meta'][doc_id]['doi']
    temp_info_dict['year'] = data['meta'][doc_id]['year']
    
    doc_info_dict_list.append(temp_info_dict)
    
    doc_id +=1


# Collect data in a class object

class Paper(object):
    def __init__(self,
                 doi,
                 authors,
                 year,
                 full_abstract,
                 processed_abstract,
                 abstract_word_distr):
        
        self.doi = doi
        self.authors = authors
        self.year = year
        self.full_abstract = full_abstract
        self.processed_abstract = processed_abstract
        self.abstract_word_distr = abstract_word_distr

class APSCorpus(object):
    def __init__(self,
                 vocab,
                 papers_info_dict_list,
                 process = True):
        
        self.vocab = vocab
        self.id_to_word = defaultdict()
        self.word_to_id = defaultdict()
        self.papers = []
        self.doi_to_paperidx = defaultdict()
        
        if process:
            self.get_id_word_mapping()
            self.populate_papers(papers_info_dict_list)
            
    def get_id_word_mapping(self):
        # vocab_id to actual word mapping
        for idx,word in enumerate(self.vocab):
            self.id_to_word[idx] = word
            self.word_to_id[word] = idx
            
    def populate_papers(self,papers_info_dict_list):
        for paper_dict in papers_info_dict_list:
            self.doi_to_paperidx[paper_dict['doi']] = len(self.papers)
            self.papers.append(Paper(paper_dict['doi'],
                                     None,
                                     paper_dict['year'],
                                     paper_dict['full_abstract'],
                                     paper_dict['updated_abstract'],
                                     paper_dict['abstract_word_distr']
                                    ))

corpusobj = APSCorpus(vocab = data['vocab'],papers_info_dict_list = doc_info_dict_list)

# Save data to disk

# Save data to file 
file_aps = open('data/apscorpus.obj', 'wb') 
pickle.dump(corpusobj, file_aps)

# load data
filehandler = open('data/apscorpus.obj', 'rb') 
data_obj = pickle.load(filehandler)

# Example use case
id_ = 249000
data_obj.papers[id_].processed_abstract

data_obj.papers[id_].full_abstract

data_obj.papers[id_].doi

data_obj.papers[id_].year

data_obj.doi_to_paperidx['10.1103/PhysRevLett.61.1768']

# Draft

data.keys()

len(data['meta'])

data['meta'][0].keys()

data['meta'][1]['doi']

len(data['vocab'])

data['vocab'][84083]

len(data['documents'])

data['documents']['1']

# vocab_id to actual word mapping
id_to_word = defaultdict()
word_to_id = defaultdict()
for idx,word in enumerate(data['vocab']):
    id_to_word[idx] = word
    word_to_id[word] = idx

len(id_to_word.keys())

min(id_to_word.keys()),max(id_to_word.keys())

# documentwise: convert <id to count> mapping to <word to count> mapping
doc_word_count_dict = defaultdict()

doc_id = 0
for doc_id_str, doc_list in data['documents'].items():
    doc_dict = defaultdict()
    for id_,count_ in zip(doc_list[0],doc_list[1]):
        doc_dict[id_to_word[id_-1]]=count_ # -1 to compensate for R being 1-indexed
    doc_word_count_dict[doc_id] = copy.deepcopy(doc_dict)
    doc_id +=1
    

doc_word_count_dict[1]

abstracts_list[1]

len(abstracts_list)

# these are the ids in processed$docs.removed and out$docs.removed during preproecessing in R
docs_removed = [4920,10229,16848,17604,24092,68898,104301,108465,109275,118546,118907,128414,132047,135300,
                157901,159166,159813,170490,184016,191411,195803,203451,204141,214446,221736,223792,241881,
                246810,88613,126377] 

abstracts_list[24091]

len(abstracts_list[4919])

updated_abstracts_list = []
for idx,abstract in enumerate(abstracts_list):
    if (idx+1 in docs_removed):
        pass
    else:
        updated_abstracts_list.append(abstract)

len(updated_abstracts_list)
