==> Raiyan_code/4b_export_frex_to_json.ipynb <==
from google.colab import drive
drive.mount('/content/gdrive')

cd gdrive/MyDrive/Colab\ Notebooks/APS2/

!pip install pyreadr
import pyreadr
import numpy as np
import pandas as pd
from collections import defaultdict
import json 

for k in [20,28,30,40,50,60]:
    for frex in [0.25,0.5,0.75]:
        filename = "frexconcepts_k"+str(k)+"_frex"+str(frex)
        in_file = pyreadr.read_r(filename+".rda")
        df = in_file['a'] # 'a' was the variable name of the object when exporting from R
        out_dict = defaultdict()
        for row_idx in range(k):
            out_dict[row_idx] = list(df.iloc[row_idx])
        with open(filename+".json", "w") as outfile:
            json.dump(out_dict, outfile)
