==> Raiyan_code/5_word2vec_prep.ipynb <==
import numpy as np
import pandas as pd 
import json
from collections import defaultdict
import copy
import pickle 

# Load APS corpus

class Paper(object):
    def __init__(self,
                 doi,
                 authors,
                 year,
                 full_abstract,
                 processed_abstract,
                 abstract_word_distr):
        
        self.doi = doi
        self.authors = authors
        self.year = year
        self.full_abstract = full_abstract
        self.processed_abstract = processed_abstract
        self.abstract_word_distr = abstract_word_distr

class APSCorpus(object):
    def __init__(self,
                 vocab,
                 papers_info_dict_list,
                 process = True):
        
        self.vocab = vocab
        self.id_to_word = defaultdict()
        self.word_to_id = defaultdict()
        self.papers = []
        self.doi_to_paperidx = defaultdict()
        
        if process:
            self.get_id_word_mapping()
            self.populate_papers(papers_info_dict_list)
            
    def get_id_word_mapping(self):
        # vocab_id to actual word mapping
        for idx,word in enumerate(self.vocab):
            self.id_to_word[idx] = word
            self.word_to_id[word] = idx
            
    def populate_papers(self,papers_info_dict_list):
        for paper_dict in papers_info_dict_list:
            self.doi_to_paperidx[paper_dict['doi']] = len(self.papers)
            self.papers.append(Paper(paper_dict['doi'],
                                     None,
                                     paper_dict['year'],
                                     paper_dict['full_abstract'],
                                     paper_dict['updated_abstract'],
                                     paper_dict['abstract_word_distr']
                                    ))

filehandler = open('data/apscorpus.obj', 'rb') 
corpus = pickle.load(filehandler)

id_ = 5
corpus.papers[id_].full_abstract

corpus.papers[id_].processed_abstract

# Load FREX data

klist = [20,28,30,40,50,60]
frexlist = [0.25,0.5,0.75]

frex_concepts_dict = {key1:{key2:dict() for key2 in frexlist} for key1 in klist}

for k in klist:
    for frex in frexlist:
        filename = "data/frexconcepts_k"+str(k)+"_frex"+str(frex)+".json"
        f = open(filename)
        temp_dict = json.load(f)
        frex_concepts_dict[k][frex] = copy.deepcopy(temp_dict) # [20][0.75] holds the FREX concepts for k=20, FR=0.75

# Create FREX space

frex_space_dict = {key1:{key2:list() for key2 in frexlist} for key1 in klist}

for k in klist:
    for frex in frexlist:
        temp_frexspace = list(set([v for key,val in frex_concepts_dict[k][frex].items() for v in val]))
        frex_space_dict[k][frex] = copy.deepcopy(temp_frexspace)

# Export abstracts with only the FREX terms to disk

out_dict = {key1:{key2:list() for key2 in frexlist} for key1 in klist}
full_length_list = []
filtered_length_dict = {key1:{key2:list() for key2 in frexlist} for key1 in klist}

pid=0
for paper_id in range(len(corpus.papers)):
    current_abstract = corpus.papers[paper_id].processed_abstract
    temp_split_abs = current_abstract.split()
    full_length_list.append(len(temp_split_abs))
    
    #generate output for every (k,frex) combo
    for k in klist:
        for frex in frexlist:
            temp_refined_abs = [term for term in temp_split_abs if term in frex_space_dict[k][frex]]
            temp_out_abs = ' '.join(temp_refined_abs)
            out_dict[k][frex].append(temp_out_abs)
            filtered_length_dict[k][frex].append(len(temp_refined_abs))
            
    pid=pid+1
    if pid%1000==0:
        print(pid)

for k in klist:
    for frex in frexlist:
        filename = "data/word2vec_inputs/abstracts_k"+str(k)+"_frex"+str(frex)+".txt"
        f = open(filename, "w")
        for abstr in out_dict[k][frex]:
            f.write(abstr + "\n")
        f.close()

for k in klist:
    for frex in frexlist:
        print("retained after frex filtering for ",k,frex,": ",
              np.mean(filtered_length_dict[k][frex])/np.mean(full_length_list)*100,"%")

# Draft

f = open("data/frexconcepts_k40_frex0.5.json")
frexconcepts_k40_frex50 = json.load(f)

len(frexconcepts_k40_frex50.keys()), len(frexconcepts_k40_frex50['0'])

print("Total concepts: "+str(len(frexconcepts_k40_frex50.keys())* len(frexconcepts_k40_frex50['0'])))

frexspace = list(set([v for key,val in frexconcepts_k40_frex50.items() for v in val]))

len(frexspace)

frexspace[:10]

len(corpus.papers)

len(corpus.papers[0].processed_abstract.split())

split_abs = corpus.papers[id_].processed_abstract.split()
print(split_abs)

refined_abs = [term for term in split_abs if term in frexspace]
refined_abs2 = ' '.join(refined_abs)
refined_abs2
