==> Raiyan_code/7_PMI_calculation.ipynb <==
import numpy as np
import pandas as pd
from collections import defaultdict
from itertools import combinations
import math
import copy

# Do the counting

# Generator class for RAM friendliness
class MySentences(object):
    def __init__(self, filename):
        self.file=filename
 
    def __iter__(self):
        f = open(self.file)
        for line in f:
            yield line.split()

# iterator
abstracts = MySentences('data/word2vec_inputs/abstracts_k40_frex0.5.txt')
abstract_iterator = iter(abstracts)

concept_counter = defaultdict(lambda:0)
joint_counter = defaultdict(lambda:0)
N_counter = 0

for current_abstract in abstract_iterator:
    concepts = set(current_abstract)
    for concept in concepts:
        concept_counter[concept]+=1
    for subset in combinations(list(concepts), 2):
        if (subset[1],subset[0]) not in joint_counter.keys():
            joint_counter[subset]+=1
        else:
            joint_counter[(subset[1],subset[0])]+=1
    N_counter+=1
    if N_counter%1000==0:
        print(N_counter)

print("Total abstracts, N =",N_counter)
print("Total concepts =",len(concept_counter.keys()))
print("Possible combination pairs =",math.comb(len(concept_counter.keys()), 2))
print("Total concept-pairs that co-appear =",len(joint_counter.keys()))

# PMI calculation

PMI_dict = {}

for key,val in joint_counter.items():
    if concept_counter[key[0]]>=5 and concept_counter[key[1]]>=5:
        out = (val*N_counter)/(concept_counter[key[0]]*concept_counter[key[1]])
        PMI_dict[key] = math.log10(out)

list(PMI_dict.values())[:10]

len(PMI_dict.keys())

sorted_list = sorted(PMI_dict, key=PMI_dict.get, reverse=True)

sorted_list 

len(sorted_list[:5000000])

PMI_dict[('quark', 'film')]

PMI_dict[('quark', 'puls')]

PMI_dict[('albumin', 'serum')]

PMI_dict[('heterometal', 'homometal')]

PMI_dict[('covid-', 'pandem')]

# load pickle module
import pickle


# create a binary pickle file 
f = open("data/pmi_k40_frex0.5.pkl","wb")

# write the python object (dict) to pickle file
pickle.dump(PMI_dict,f)

# close file
f.close()
